requirements.txt
fastapi
uvicorn
openai
python-dotenv
PyPDF2
python-docx
pytesseract
pywin32
pillow
pydantic
aiofiles
requests
logging
datetime
typing
python-multipart
tensorflow  
numpy  
scikit-learn  